+++ toolium/behave/env_utils.py
@@ -18,7 +18,7 @@ limitations under the License.
 
 import warnings
 import sys
-
+from pkg_resources import parse_version
 
 # constants
 # pre-actions in feature files
@@ -64,7 +64,7 @@ class Logger:
               '           - Exception: %s' % exc
         if self.logger is not None:
             self.logger.error(msg)
-        raise Exception('      ERROR - %s' % msg)
+        self.by_console('      ERROR - %s' % msg)
 
     def debug(self, value):
         """
@@ -195,7 +195,8 @@ class DynamicEnvironment:
                 self.logger.by_console('\n')
                 if action == ACTIONS_BEFORE_SCENARIO:
                     self.scenario_counter += 1
-                    self.logger.by_console("  ------------------ Scenario Nº: %d ------------------" % self.scenario_counter)
+                    self.logger.by_console(
+                        "  ------------------ Scenario Nº: %d ------------------" % self.scenario_counter)
                 self.logger.by_console('  %s:' % action)
             for item in self.actions[action]:
                 self.scenario_error = False
@@ -209,6 +210,18 @@ class DynamicEnvironment:
                     elif action in [ACTIONS_BEFORE_SCENARIO]:
                         self.scenario_error = True
                     self.logger.error(exc)
+                    break
+
+    def reset_error_status(self):
+        """
+        Check if the dyn_env has got any exception when executing the steps and restore the value of status to False.
+        :return: True if any exception has been raised when executing steps
+        """
+        try:
+            return self.feature_error or self.scenario_error
+        finally:
+            self.feature_error = False
+            self.scenario_error = False
 
     def execute_before_feature_steps(self, context):
         """
@@ -216,6 +229,10 @@ class DynamicEnvironment:
         :param context: It’s a clever place where you and behave can store information to share around, automatically managed by behave.
         """
         self.__execute_steps_by_action(context, ACTIONS_BEFORE_FEATURE)
+
+        if context.dyn_env.feature_error:
+            # Mark this Feature as skipped. Steps will not be executed.
+            context.feature.mark_skipped()
 
     def execute_before_scenario_steps(self, context):
         """
@@ -225,6 +242,10 @@ class DynamicEnvironment:
         if not self.feature_error:
             self.__execute_steps_by_action(context, ACTIONS_BEFORE_SCENARIO)
 
+        if context.dyn_env.scenario_error:
+            # Mark this Scenario as skipped. Steps will not be executed.
+            context.scenario.mark_skipped()
+
     def execute_after_scenario_steps(self, context):
         """
         actions after each scenario
@@ -233,6 +254,11 @@ class DynamicEnvironment:
         if not self.feature_error and not self.scenario_error:
             self.__execute_steps_by_action(context, ACTIONS_AFTER_SCENARIO)
 
+        # Behave dynamic environment: Fail all steps if dyn_env has got any error and reset it
+        if self.reset_error_status():
+            context.scenario.reset()
+            context.dyn_env.fail_first_step_precondition_exception(context.scenario)
+
     def execute_after_feature_steps(self, context):
         """
         actions after the feature
@@ -240,3 +266,30 @@ class DynamicEnvironment:
         """
         if not self.feature_error:
             self.__execute_steps_by_action(context, ACTIONS_AFTER_FEATURE)
+
+        # Behave dynamic environment: Fail all steps if dyn_env has got any error and reset it
+        if self.reset_error_status():
+            context.feature.reset()
+            for scenario in context.feature.walk_scenarios():
+                context.dyn_env.fail_first_step_precondition_exception(scenario)
+
+    def fail_first_step_precondition_exception(self, scenario):
+        """
+        Fail first step in the given Scenario and add exception message for the output.
+        This is needed because xUnit exporter in Behave fails if there are not failed steps.
+        :param scenario: Behave's Scenario
+        """
+
+        try:
+            import behave
+            if parse_version(behave.__version__) < parse_version('1.2.6'):
+                status = 'failed'
+            else:
+                status = behave.model_core.Status.failed
+        except ImportError as exc:
+            self.logger.error(exc)
+            raise
+
+        scenario.steps[0].status = status
+        scenario.steps[0].exception = Exception("Preconditions failed")
+        scenario.steps[0].error_message = "Failing steps due to precondition exceptions"