+++ toolium/driver_wrappers_pool.py
@@ -166,7 +166,8 @@ class DriverWrappersPool(object):
             try:
                 driver_wrapper.driver.quit()
             except Exception as e:
-                driver_wrapper.logger.warn("Capture exceptions to avoid errors in teardown method due to session timeouts: \n %s" % e)
+                driver_wrapper.logger.warn(
+                    "Capture exceptions to avoid errors in teardown method due to session timeouts: \n %s" % e)
 
     @classmethod
     def download_videos(cls, name, test_passed=True, maintain_default=False):
@@ -189,16 +190,17 @@ class DriverWrappersPool(object):
                 # Download video if necessary (error case or enabled video)
                 if (not test_passed or driver_wrapper.config.getboolean_optional('Server', 'video_enabled', False)) \
                         and driver_wrapper.remote_node_video_enabled:
-                    if driver_wrapper.server_type == 'ggr':
-                        # Download video from GGR
+                    if driver_wrapper.server_type in ['ggr', 'selenoid']:
                         name = get_valid_filename(video_name.format(name, driver_index))
                         Selenoid(driver_wrapper).download_session_video(name)
                     elif driver_wrapper.server_type == 'grid':
                         # Download video from Grid Extras
-                        driver_wrapper.utils.download_remote_video(driver_wrapper.remote_node, driver_wrapper.session_id,
-                                                               video_name.format(name, driver_index))
-            except Exception as e:
-                driver_wrapper.logger.warn("Capture exceptions to avoid errors in teardown method due to session timeouts: \n %s" % e)
+                        driver_wrapper.utils.download_remote_video(driver_wrapper.remote_node,
+                                                                   driver_wrapper.session_id,
+                                                                   video_name.format(name, driver_index))
+            except Exception as exc:
+                # Capture exceptions to avoid errors in teardown method due to session timeouts
+                driver_wrapper.logger.warn('Error downloading videos: %s' % exc)
             driver_index += 1
 
     @classmethod
@@ -219,14 +221,14 @@ class DriverWrappersPool(object):
         log_name = '{} [driver {}]' if len(cls.driver_wrappers) > 1 else '{}'
         driver_index = 1
         for driver_wrapper in cls.driver_wrappers:
-            if not driver_wrapper.driver:
+            if not driver_wrapper.driver or driver_wrapper.server_type in ['ggr', 'selenoid']:
                 continue
             if driver_wrapper.config.getboolean_optional('Server', 'logs_enabled') or not test_passed:
                 try:
                     driver_wrapper.utils.save_webdriver_logs(log_name.format(test_name, driver_index))
-                except Exception:
+                except Exception as exc:
                     # Capture exceptions to avoid errors in teardown method due to session timeouts
-                    pass
+                    driver_wrapper.logger.warn('Error downloading webdriver logs: %s' % exc)
             driver_index += 1
 
     @classmethod
@@ -239,15 +241,15 @@ class DriverWrappersPool(object):
         log_name = '{} [driver {}]' if len(cls.driver_wrappers) > 1 else '{}'
         driver_index = 1
         for driver_wrapper in cls.driver_wrappers:
-            if not driver_wrapper.driver or driver_wrapper.server_type != 'ggr':
+            if not driver_wrapper.driver or driver_wrapper.server_type not in ['ggr', 'selenoid']:
                 continue
             try:
                 if driver_wrapper.config.getboolean_optional('Server', 'logs_enabled') or not test_passed:
                     name = get_valid_filename(log_name.format(test_name, driver_index))
                     Selenoid(driver_wrapper).download_session_log(name)
-            except Exception:
+            except Exception as exc:
                 # Capture exceptions to avoid errors in teardown method due to session timeouts
-                pass
+                driver_wrapper.logger.warn('Error downloading GGR logs: %s' % exc)
             driver_index += 1
 
     @staticmethod