+++ toolium/selenoid.py
@@ -23,6 +23,7 @@ import requests
 
 # constants
 STATUS_OK = 200
+STATUS_PORT = "8888"
 DOWNLOADS_PATH = u'downloads'
 MP4_EXTENSION = u'mp4'
 LOG_EXTENSION = u'log'
@@ -50,31 +51,26 @@ class Selenoid(object):
        the files are always removed in the selenoid server
     """
 
-    def __init__(self, driver_wrapper, hub=True, **kwargs):
+    def __init__(self, driver_wrapper, **kwargs):
         """
         get data from properties file and session
         :param driver_wrapper: driver_wrapper instance from toolium
-        :param hub: determine if the server is a node (Selenoid-False) or a hub (GGR-True)
         dynamic parameters:
         :param videos_dir: videos directory to download
         :param logs_dir: logs directory to download
         :param output_dir: general directory to download
         """
         self.driver_wrapper = driver_wrapper
-        self.hub = hub
         from toolium.driver_wrappers_pool import DriverWrappersPool
         self.videos_directory = kwargs.get('videos_dir', DriverWrappersPool.videos_directory)
         self.logs_directory = kwargs.get('logs_dir', DriverWrappersPool.logs_directory)
         self.output_directory = kwargs.get('output_dir', DriverWrappersPool.output_directory)
-
-        driver_wrapper.logger.info('Selenoid download is beginning, using GGR as hub: {}...'.format(self.hub))
         self.browser_remote = driver_wrapper.config.getboolean_optional('Server', 'enabled', False)
-        self.enabled_logs = driver_wrapper.config.getboolean_optional('Server', 'logs_enabled', False)
+        self.browser = driver_wrapper.driver.desired_capabilities['browserName']
 
         if self.browser_remote:
             self.session_id = driver_wrapper.driver.session_id
             self.server_url = driver_wrapper.utils.get_server_url()
-            driver_wrapper.logger.info('Selenoid host info: \n %s' % self.get_selenoid_info())
 
     def __download_file(self, url, path_file, timeout):
         """
@@ -86,10 +82,13 @@ class Selenoid(object):
         """
         status_code = 0
         init_time = time.time()
+        self.driver_wrapper.logger.info('Downloading file from Selenoid node: %s' % url)
         # retries policy
         while status_code != STATUS_OK and time.time() - init_time < float(timeout):
             body = requests.get(url)
             status_code = body.status_code
+            if status_code != STATUS_OK:
+                time.sleep(1)
         took = time.time() - init_time  # time used to download the file
         # create the folders and store the file downloaded
         if status_code == STATUS_OK:
@@ -100,14 +99,13 @@ class Selenoid(object):
                 fp = open(path_file, 'wb')
                 fp.write(body.content)
                 fp.close()
-                self.driver_wrapper.logger.info('the %s file has been downloaded successfully and took %d'
-                                                ' seconds' % (path_file, took))
+                self.driver_wrapper.logger.info('File has been downloaded successfully to "%s" and took %d '
+                                                'seconds' % (path_file, took))
                 return True
             except IOError as e:
-                self.driver_wrapper.logger.warn('the %s file has a problem; \n %s' % (path_file, e))
+                self.driver_wrapper.logger.warn('Error writing downloaded file in "%s":\n %s' % (path_file, e))
         else:
-            self.driver_wrapper.logger.warn('the file to download does not exist in the server after %s seconds'
-                                            ' (timeout).' % timeout)
+            self.driver_wrapper.logger.warn('File "%s" does not exist in the server after %s seconds' % (url, timeout))
         return False
 
     def __remove_file(self, url):
@@ -118,21 +116,42 @@ class Selenoid(object):
 
     def get_selenoid_info(self):
         """
-        retrieve the current selenoid host info. Used only in hub mode (GGR)
+        retrieve the current selenoid host info
         request: http://<username>:<password>@<ggr_host>:<ggr_port>/host/<ggr_session_id>
         :return: dict
         """
-        if self.hub:
-            host_url = '{}/host/{}'.format(self.server_url, self.session_id)
-            try:
-                response = requests.get(host_url)
-            except Exception as e:
-                self.driver_wrapper.logger.warn("the GGR request to get the node host is failed: \n  %s" % e)
-                return None
-            return response.json()
-        return None
+        host_url = '{}/host/{}'.format(self.server_url, self.session_id)
+        try:
+            selenoid_info = requests.get(host_url).json()
+        except Exception:
+            return None
+        self.driver_wrapper.logger.info('Selenoid host info: \n %s' % selenoid_info)
+        return selenoid_info
 
-    def download_session_video(self, scenario_name, timeout=10):
+    def is_the_session_still_active(self):
+        """
+        Is the GGR session still active? Associated to a browser and the sessionId
+        Example of GGR status:
+        {"browsers":{"MicrosoftEdge":{"latest":{}},"android":{"8.1":{}},"chrome":{"70.0":{},"latest":{"test_tef":{"count":1,"sessions":[{"caps":{"browserName":"chrome","enableVNC":true,"enableVideo":true,"platform":"ANY","screenResolution":"1280x1024x24","version":"latest","videoName":"selenoide952e551bb9395e16d060f28c54e5d31.mp4","videoScreenSize":"1280x1024"},"container":"8489205e28c9781472e99c3921a6240de3894a3603ed9e187ad6360b6b013b8b","containerInfo":{"id":"8489205e28c9781472e99c3921a6240de3894a3603ed9e187ad6360b6b013b8b","ip":"172.17.0.4"},"id":"1345506093dfed8dbcef610da476911a228ca315978e5464ae49fb1142bbc49b","screen":"1280x1024x24","vnc":true}]}}},"firefox":{"59.0":{},"63.0":{},"64.0":{},"latest":{}},"internet explorer":{"11":{}},"opera":{"56.0":{},"latest":{}},"safari":{"latest":{}}},"pending":0,"queued":0,"total":30,"used":1}
+        :return boolean (although in case of error in the request will be returned None)
+        """
+        server_url_splitted = self.server_url.split(":")
+        host_url = "{}:{}:{}:{}/status".format(server_url_splitted[0], server_url_splitted[1], server_url_splitted[2], STATUS_PORT)
+
+        try:
+            response = requests.get(host_url).json()["browsers"][self.browser]
+        except Exception as e:
+            self.driver_wrapper.logger.warn("the GGR status request has failed: \nResponse:  %s \nError message: %s\n" % (response.content, e))
+            return None
+        for browser in response:
+            if response[browser] != {}:
+                sessions = response[browser][server_url_splitted[1].split("@")[0].replace("//", "")]["sessions"]
+                for session in sessions:
+                    if session["id"] == self.session_id:
+                        return True
+        return False
+
+    def download_session_video(self, scenario_name, timeout=5):
         """
         download the execution video file if the scenario fails or the video is enabled,
         renaming the file to scenario name and removing the video file in the server.
@@ -141,19 +160,24 @@ class Selenoid(object):
         :param scenario_name: scenario name
         :param timeout: threshold until the video file is downloaded
         """
+        # Download video only in linux nodes with video enabled
+        if (self.driver_wrapper.get_driver_platform().lower() != 'linux' or
+                not self.driver_wrapper.config.getboolean_optional('Capabilities', 'enableVideo')):
+            return
+
         path_file = os.path.join(self.videos_directory, '%s.%s' % (scenario_name, MP4_EXTENSION))
-        if not self.hub:
+        if self.driver_wrapper.server_type == 'selenoid':
             filename = '%s.%s' % (self.session_id, MP4_EXTENSION)
         else:
             filename = self.session_id
         video_url = '{}/video/{}'.format(self.server_url, filename)
-        # download the execution video file if the scenario is failed
+        # download the execution video file
         if self.browser_remote:
             self.__download_file(video_url, path_file, timeout)
         # remove the video file if it does exist
         self.__remove_file(video_url)
 
-    def download_session_log(self, scenario_name, timeout=10):
+    def download_session_log(self, scenario_name, timeout=5):
         """
         download the session log file from remote selenoid,
         renaming the file to scenario name and removing the video file in the server.
@@ -162,19 +186,24 @@ class Selenoid(object):
         :param scenario_name: scenario name
         :param timeout: threshold until the video file is downloaded
         """
+        # Download logs only in linux nodes with logs enabled
+        if (self.driver_wrapper.get_driver_platform().lower() != 'linux' or
+                not self.driver_wrapper.config.getboolean_optional('Capabilities', 'enableLog')):
+            return
+
         path_file = os.path.join(self.logs_directory, '%s_ggr.%s' % (scenario_name, LOG_EXTENSION))
-        if not self.hub:
+        if self.driver_wrapper.server_type == 'selenoid':
             filename = '%s.%s' % (self.session_id, LOG_EXTENSION)
         else:
             filename = self.session_id
         logs_url = '{}/logs/{}'.format(self.server_url, filename)
         # download the session log file
-        if self.enabled_logs and self.browser_remote:
+        if self.browser_remote:
             self.__download_file(logs_url, path_file, timeout)
         # remove the log file if it does exist
         self.__remove_file(logs_url)
 
-    def download_file(self, filename, timeout=10):
+    def download_file(self, filename, timeout=5):
         """
         download a file from remote selenoid and removing the file in the server.
         request: http://<username>:<password>@<ggr_host>:<ggr_port>/download/<ggr_session_id>/<filename>
@@ -184,7 +213,7 @@ class Selenoid(object):
         """
         path_file = os.path.join(self.output_directory, DOWNLOADS_PATH, self.session_id[-8:], filename)
         file_url = '{}/download/{}/{}'.format(self.server_url, self.session_id, filename)
-        # download the session log file
+        # download the file
         if self.browser_remote:
             self.__download_file(file_url, path_file, timeout)
             return path_file